Moved native build code to NativeBuild.scala.
