+++ project/scalatest.scala
@@ -15,11 +15,6 @@ import org.scalajs.sbtplugin.ScalaJSPlugin.autoImport.{scalaJSLinkerConfig, jsEn
 import org.portablescala.sbtplatformdeps.PlatformDepsPlugin.autoImport._
 import sbtcrossproject.CrossPlugin.autoImport.{crossProject, CrossType, _}
 import scalajscrossproject.ScalaJSCrossPlugin.autoImport._
-import scalanativecrossproject.ScalaNativeCrossPlugin.autoImport._
-
-import scalanative.sbtplugin.ScalaNativePlugin
-import scalanative.sbtplugin.ScalaNativePluginInternal.NativeTest
-import ScalaNativePlugin.autoImport._
 
 import com.typesafe.tools.mima.plugin.MimaKeys.{mimaPreviousArtifacts, mimaCurrentClassfiles, mimaBinaryIssueFilters}
 import com.typesafe.tools.mima.core._
@@ -27,7 +22,7 @@ import com.typesafe.tools.mima.core.ProblemFilters._
 
 import dotty.tools.sbtplugin.DottyPlugin.autoImport._
 
-object ScalatestBuild extends BuildCommons with DottyBuild {
+object ScalatestBuild extends BuildCommons with DottyBuild with NativeBuild {
 
   // To run gentests
   // rm -rf gentests
@@ -183,20 +178,6 @@ object ScalatestBuild extends BuildCommons with DottyBuild {
         Seq.empty
     }
 
-
-
-  lazy val nativeCrossBuildLibraryDependencies = Def.setting {
-    CrossVersion.partialVersion(scalaVersion.value) match {
-      // if scala 2.11+ is used, add dependency on scala-xml module
-      case Some((2, scalaMajor)) if scalaMajor >= 11 =>
-        Seq(
-          "org.scala-lang.modules" %% "scala-xml" % "1.0.6"
-        )
-      case _ =>
-        Seq.empty
-    }
-  }
-
   def scalaLibraries(theScalaVersion: String) =
     Seq(
       "org.scala-lang" % "scala-compiler" % theScalaVersion % "provided",
@@ -305,35 +286,6 @@ object ScalatestBuild extends BuildCommons with DottyBuild {
       "-m", "org.scalatest.diagrams",
       "-oDIF"))
 
-  def scalatestTestNativeOptions =
-    Seq(Tests.Argument(TestFrameworks.ScalaTest,
-      "-l", "org.scalatest.tags.Slow",
-      "-m", "org.scalatest",
-      "-m", "org.scalactic",
-      "-m", "org.scalactic.anyvals",
-      "-m", "org.scalactic.algebra",
-      "-m", "org.scalactic.enablers",
-      "-m", "org.scalatest.fixture",
-      "-m", "org.scalatest.concurrent",
-      "-m", "org.scalatest.events",
-      "-m", "org.scalatest.prop",
-      "-m", "org.scalatest.tools",
-      "-m", "org.scalatest.matchers",
-      "-m", "org.scalatest.matchers",
-      "-m", "org.scalatest.matchers.should",
-      "-m", "org.scalatest.matchers.must",
-      "-m", "org.scalatest.matchers.dsl",
-      "-m", "org.scalatest.verbs",
-      "-m", "org.scalatest.suiteprop",
-      "-m", "org.scalatest.path",
-      "-m", "org.scalatest.exceptions",
-      "-m", "org.scalatest.time",
-      "-m", "org.scalatest.words",
-      "-m", "org.scalatest.enablers",
-      "-m", "org.scalatest.expectations",
-      "-m", "org.scalatest.diagrams",
-      "-oDIF"))
-
   lazy val commonTest = Project("common-test", file("common-test"))
     .settings(sharedSettings: _*)
     .settings(
@@ -368,23 +320,6 @@ object ScalatestBuild extends BuildCommons with DottyBuild {
       publishLocal := {},
       scalacOptions in (Compile, doc) := List.empty
     ).dependsOn(scalacticMacroJS, LocalProject("scalatestJS")).enablePlugins(ScalaJSPlugin)
-
-    lazy val commonTestNative = Project("commonTestNative", file("common-test.native"))
-      .settings(sharedSettings: _*)
-      .settings(
-        projectTitle := "Common test classes used by scalactic.native and scalatest.native",
-        sourceGenerators in Compile += {
-          Def.task{
-            GenCommonTestNative.genMain((sourceManaged in Compile).value / "scala" / "org" / "scalatest", version.value, scalaVersion.value) ++
-            GenGen.genMain((sourceManaged in Compile).value / "scala" / "org" / "scalatest" / "prop", version.value, scalaVersion.value) ++
-            GenCompatibleClasses.genTest((sourceManaged in Compile).value, version.value, scalaVersion.value)
-          }.taskValue
-        },
-        publishArtifact := false,
-        publish := {},
-        publishLocal := {},
-        scalacOptions in (Compile, doc) := List.empty
-      ).dependsOn(scalacticMacroNative, LocalProject("scalatestNative")).enablePlugins(ScalaNativePlugin)
 
   lazy val scalacticMacro = Project("scalacticMacro", file("scalactic-macro"))
     .settings(sharedSettings: _*)
@@ -436,27 +371,6 @@ object ScalatestBuild extends BuildCommons with DottyBuild {
       }.triggeredBy(compile in Compile).value,
       scalacOptions in (Compile, doc) := List.empty
     ).enablePlugins(ScalaJSPlugin)
-
-  lazy val scalacticMacroNative = Project("scalacticMacroNative", file("scalactic-macro.native"))
-    .settings(sharedSettings: _*)
-    .settings(
-      projectTitle := "Scalactic Macro.native",
-      organization := "org.scalactic",
-      sourceGenerators in Compile += {
-        Def.task{
-          GenScalacticNative.genMacroScala((sourceManaged in Compile).value / "scala", version.value, scalaVersion.value) ++
-            ScalacticGenResourcesJSVM.genResources((sourceManaged in Compile).value / "scala" / "org" / "scalactic", version.value, scalaVersion.value) ++
-            GenAnyVals.genMain((sourceManaged in Compile).value / "scala" / "org" / "scalactic" / "anyvals", version.value, scalaVersion.value, false) ++
-            GenEvery.genMain((sourceManaged in Compile).value / "org" / "scalactic", version.value, scalaVersion.value) ++
-            GenColCompatHelper.genMain((sourceManaged in Compile).value / "org" / "scalactic", version.value, scalaVersion.value)
-        }.taskValue
-      },
-      // Disable publishing macros directly, included in scalactic main jar
-      publishArtifact := false,
-      publish := {},
-      publishLocal := {},
-      scalacOptions in (Compile, doc) := List.empty
-    ).enablePlugins(ScalaNativePlugin)
 
   lazy val scalactic = Project("scalactic", file("scalactic"))
     .enablePlugins(SbtOsgi)
@@ -556,49 +470,6 @@ object ScalatestBuild extends BuildCommons with DottyBuild {
       )
     ).dependsOn(scalacticMacroJS % "compile-internal, test-internal").enablePlugins(ScalaJSPlugin)
 
-  lazy val scalacticNative = Project("scalacticNative", file("scalactic.native"))
-    .enablePlugins(SbtOsgi)
-    .settings(sharedSettings: _*)
-    .settings(scalacticDocSettings: _*)
-    .settings(
-      projectTitle := "Scalactic.native",
-      organization := "org.scalactic",
-      moduleName := "scalactic",
-      sourceGenerators in Compile += {
-        Def.task {
-          GenScalacticNative.genScala((sourceManaged in Compile).value / "scala", version.value, scalaVersion.value) ++
-          ScalacticGenResourcesJSVM.genFailureMessages((sourceManaged in Compile).value / "scala", version.value, scalaVersion.value) ++
-          GenArrayHelper.genMain((sourceManaged in Compile).value / "org" / "scalactic", version.value, scalaVersion.value)
-        }.taskValue
-      },
-      resourceGenerators in Compile += {
-        Def.task {
-          GenScalacticJS.genResource((sourceManaged in Compile).value / "scala", version.value, scalaVersion.value)
-        }.taskValue
-      }
-    ).settings(osgiSettings: _*).settings(
-    OsgiKeys.exportPackage := Seq(
-      "org.scalactic",
-      "org.scalactic.anyvals",
-      "org.scalactic.exceptions",
-      "org.scalactic.source"
-    ),
-    OsgiKeys.importPackage := Seq(
-      "org.scalatest.*",
-      "org.scalactic.*",
-      "scala.util.parsing.*;version=\"$<range;[==,=+);$<replace;1.0.4;-;.>>\"",
-      "scala.xml.*;version=\"$<range;[==,=+);$<replace;1.0.4;-;.>>\"",
-      "scala.*;version=\"$<range;[==,=+);$<replace;"+scalaBinaryVersion.value+";-;.>>\"",
-      "*;resolution:=optional"
-    ),
-    OsgiKeys.additionalHeaders:= Map(
-      "Bundle-Name" -> "Scalactic",
-      "Bundle-Description" -> "Scalactic.js is an open-source library for Scala-js projects.",
-      "Bundle-DocURL" -> "http://www.scalactic.org/",
-      "Bundle-Vendor" -> "Artima, Inc."
-    )
-  ).dependsOn(scalacticMacroNative % "compile-internal, test-internal").enablePlugins(ScalaNativePlugin)
-
   lazy val scalacticTest = Project("scalactic-test", file("scalactic-test"))
     .settings(sharedSettings: _*)
     .settings(
@@ -647,28 +518,6 @@ object ScalatestBuild extends BuildCommons with DottyBuild {
       publishLocal := {},
       scalacOptions ++= (if (scalaBinaryVersion.value == "2.10" || scalaVersion.value.startsWith("2.13")) Seq.empty[String] else Seq("-Ypartial-unification"))
     ).dependsOn(scalacticJS, scalatestJS % "test", commonTestJS % "test").enablePlugins(ScalaJSPlugin)
-
-  lazy val scalacticTestNative = Project("scalacticTestNative", file("scalactic-test.native"))
-    .settings(sharedSettings: _*)
-    .settings(
-      projectTitle := "Scalactic Test.native",
-      organization := "org.scalactic",
-      testOptions in Test ++=
-        Seq(Tests.Argument(TestFrameworks.ScalaTest, "-oDIF")),
-      /*nativeOptimizerDriver in NativeTest := {
-        val orig = tools.OptimizerDriver((nativeConfig in NativeTest).value)
-        orig.withPasses(orig.passes.filterNot(p => p == pass.DeadBlockElimination || p == pass.GlobalBoxingElimination))
-      },*/
-      nativeLinkStubs in NativeTest := true,
-      sourceGenerators in Test += {
-        Def.task {
-          GenScalacticNative.genTest((sourceManaged in Test).value / "scala", version.value, scalaVersion.value)
-        }.taskValue
-      },
-      publishArtifact := false,
-      publish := {},
-      publishLocal := {}
-    ).dependsOn(scalacticNative, scalatestNative % "test", commonTestNative % "test").enablePlugins(ScalaNativePlugin)
 
   lazy val scalatest = Project("scalatest", file("scalatest"))
    .enablePlugins(SbtOsgi)
@@ -917,143 +766,6 @@ object ScalatestBuild extends BuildCommons with DottyBuild {
         }
     ).dependsOn(scalatestJS % "test", commonTestJS % "test").enablePlugins(ScalaJSPlugin)
 
-  lazy val scalatestNative = Project("scalatestNative", file("scalatest.native"))
-    .enablePlugins(SbtOsgi)
-    .settings(sharedSettings: _*)
-    .settings(scalatestDocSettings: _*)
-    .settings(
-      projectTitle := "ScalaTest",
-      organization := "org.scalatest",
-      moduleName := "scalatest",
-      initialCommands in console := """|import org.scalatest._
-                                       |import org.scalactic._
-                                       |import Matchers._""".stripMargin,
-      libraryDependencies += "org.scala-native" %%% "test-interface" % "0.4.0-M2",
-      //jsDependencies += RuntimeDOM % "test",
-      sourceGenerators in Compile += {
-        Def.task {
-          GenScalaTestNative.genHtml((resourceManaged in Compile).value, version.value, scalaVersion.value)
-
-          GenScalaTestNative.genScala((sourceManaged in Compile).value / "scala", version.value, scalaVersion.value) ++
-          GenVersions.genScalaTestVersions((sourceManaged in Compile).value / "scala" / "org" / "scalatest", version.value, scalaVersion.value) ++
-          ScalaTestGenResourcesJSVM.genResources((sourceManaged in Compile).value / "scala" / "org" / "scalatest", version.value, scalaVersion.value) ++
-          ScalaTestGenResourcesJSVM.genFailureMessages((sourceManaged in Compile).value / "scala" / "org" / "scalatest", version.value, scalaVersion.value) ++
-          GenConfigMap.genMain((sourceManaged in Compile).value / "org" / "scalatest", version.value, scalaVersion.value)
-        }.taskValue
-      },
-      javaSourceManaged := target.value / "java",
-      managedSourceDirectories in Compile += javaSourceManaged.value,
-      sourceGenerators in Compile += {
-        Def.task{
-          GenScalaTestNative.genJava((javaSourceManaged in Compile).value / "java", version.value, scalaVersion.value)
-        }.taskValue
-      },
-      resourceGenerators in Compile += {
-        Def.task {
-          GenScalaTestNative.genHtml((resourceManaged in Compile).value, version.value, scalaVersion.value)
-        }.taskValue
-      },
-      //unmanagedResourceDirectories in Compile <+= sourceManaged( _ / "resources" ),
-      sourceGenerators in Compile += {
-        Def.task{
-          GenTable.genMainForScalaJS((sourceManaged in Compile).value / "org" / "scalatest", version.value, scalaVersion.value) ++
-          GenMatchers.genMainForScalaJS((sourceManaged in Compile).value / "org" / "scalatest", version.value, scalaVersion.value) ++
-          GenFactories.genMainJS((sourceManaged in Compile).value / "org" / "scalatest" / "matchers", version.value, scalaVersion.value)
-          //GenSafeStyles.genMainForScalaJS((sourceManaged in Compile).value / "org" / "scalatest", version.value, scalaVersion.value)
-        }.taskValue
-      },
-      scalatestJSDocTaskSetting
-    ).settings(osgiSettings: _*).settings(
-    OsgiKeys.exportPackage := Seq(
-      "org.scalatest",
-      "org.scalatest.compatible",
-      "org.scalatest.concurrent",
-      "org.scalatest.check",
-      "org.scalatest.diagrams",
-      "org.scalatest.enablers",
-      "org.scalatest.events",
-      "org.scalatest.exceptions",
-      "org.scalatest.expectations",
-      "org.scalatest.fixture",
-      "org.scalatest.funsuite",
-      "org.scalatest.featurespec",
-      "org.scalatest.funspec",
-      "org.scalatest.matchers",
-      "org.scalatest.matchers.should",
-      "org.scalatest.matchers.must",
-      "org.scalatest.matchers.dsl",
-      "org.scalatest.verbs",
-      "org.scalatest.path",
-      "org.scalatest.prop",
-      "org.scalatest.propspec",
-      "org.scalatest.tags",
-      "org.scalatest.tagobjects",
-      "org.scalatest.time",
-      "org.scalatest.tools",
-      "org.scalatest.verb",
-      "org.scalatest.words", 
-      "org.scalatest.wordspec"
-    ),
-    OsgiKeys.importPackage := Seq(
-      "org.scalatest.*",
-      "org.scalactic.*",
-      "scala.util.parsing.*;version=\"$<range;[==,=+);$<replace;1.0.4;-;.>>\"",
-      "scala.xml.*;version=\"$<range;[==,=+);$<replace;1.0.4;-;.>>\"",
-      "scala.*;version=\"$<range;[==,=+);$<replace;"+scalaBinaryVersion.value+";-;.>>\"",
-      "*;resolution:=optional"
-    ),
-    OsgiKeys.additionalHeaders:= Map(
-      "Bundle-Name" -> "ScalaTest",
-      "Bundle-Description" -> "ScalaTest.js is an open-source test framework for the Javascript Platform designed to increase your productivity by letting you write fewer lines of test code that more clearly reveal your intent.",
-      "Bundle-DocURL" -> "http://www.scalatest.org/",
-      "Bundle-Vendor" -> "Artima, Inc.",
-      "Main-Class" -> "org.scalatest.tools.Runner"
-    )
-  ).dependsOn(scalacticMacroNative % "compile-internal, test-internal", scalacticNative).enablePlugins(ScalaNativePlugin)
-
-  lazy val scalatestTestNative = Project("scalatestTestNative", file("scalatest-test.native"))
-    .settings(sharedSettings: _*)
-    .settings(
-      projectTitle := "ScalaTest Test",
-      organization := "org.scalatest",
-      libraryDependencies ++= nativeCrossBuildLibraryDependencies.value,
-      // libraryDependencies += "io.circe" %%% "circe-parser" % "0.7.1" % "test",
-      fork in test := false,
-      /*nativeOptimizerDriver in NativeTest := {
-        val orig = tools.OptimizerDriver((nativeConfig in NativeTest).value)
-        orig.withPasses(orig.passes.filterNot(p => p == pass.DeadBlockElimination || p == pass.GlobalBoxingElimination))
-      },
-      nativeOptimizerReporter in NativeTest := new tools.OptimizerReporter {
-        override def onStart(batchId: Int, batchDefns: Seq[scalanative.nir.Defn]): Unit = {
-          println(s"start $batchId")
-        }
-        override def onPass(batchId: Int, passId: Int, pass: scala.scalanative.optimizer.Pass, batchDefns: Seq[scalanative.nir.Defn]): Unit = {
-          println(s"$batchId ${pass.getClass.getSimpleName}")
-        }
-        override def onComplete(batchId: Int, batchDefns: Seq[scalanative.nir.Defn]): Unit = {
-          println(s"end $batchId")
-        }
-      },*/
-      nativeLinkStubs in NativeTest := true,
-      testOptions in Test := scalatestTestNativeOptions,
-      publishArtifact := false,
-      publish := {},
-      publishLocal := {},
-      sourceGenerators in Test += {
-        Def.task {
-          GenScalaTestNative.genTest((sourceManaged in Test).value / "scala", version.value, scalaVersion.value)
-        }.taskValue
-      },
-      sourceGenerators in Test +=
-        Def.task {
-          GenGen.genTestForNative((sourceManaged in Test).value, version.value, scalaVersion.value)
-        }/*,  // OOM even with 14gb heap size, will turn this one when 32gb machine is common or if newer scala-native use lesser memory.
-      sourceGenerators in Test +=
-        Def.task {
-          GenMustMatchersTests.genTestForScalaNative((sourceManaged in Test).value, version.value, scalaVersion.value)
-        }*/
-    ).dependsOn(scalatestNative % "test", commonTestNative % "test").enablePlugins(ScalaNativePlugin)
-
   lazy val scalatestApp = Project("scalatestApp", file("."))
     .enablePlugins(SbtOsgi)
     .settings(sharedSettings: _*)
@@ -1214,84 +926,6 @@ object ScalatestBuild extends BuildCommons with DottyBuild {
         "Main-Class" -> "org.scalatest.tools.Runner"
       )
     ).dependsOn(scalacticMacroJS % "compile-internal, test-internal", scalacticJS % "compile-internal", scalatestJS % "compile-internal").aggregate(scalacticMacroJS, scalacticJS, scalatestJS, commonTestJS, scalacticTestJS, scalatestTestJS).enablePlugins(ScalaJSPlugin)
-
-    lazy val scalatestAppNative = Project("scalatestAppNative", file("scalatest-app.native"))
-      .enablePlugins(SbtOsgi)
-      .settings(sharedSettings: _*)
-      .settings(
-        projectTitle := "ScalaTest App",
-        name := "scalatest-app",
-        organization := "org.scalatest",
-        moduleName := "scalatest-app",
-        libraryDependencies ++= nativeCrossBuildLibraryDependencies.value,
-        libraryDependencies += "org.scala-native" %%% "test-interface" % "0.4.0-M2",
-        // include the scalactic classes and resources in the jar
-        mappings in (Compile, packageBin) ++= mappings.in(scalacticNative, Compile, packageBin).value,
-        // include the scalactic sources in the source jar
-        mappings in (Compile, packageSrc) ++= mappings.in(scalacticNative, Compile, packageSrc).value,
-        // include the scalatest classes and resources in the jar
-        mappings in (Compile, packageBin) ++= mappings.in(scalacticNative, Compile, packageBin).value,
-        // include the scalatest sources in the source jar
-        mappings in (Compile, packageSrc) ++= mappings.in(scalacticNative, Compile, packageSrc).value,
-        sourceGenerators in Compile += {
-          // Little trick to get rid of bnd error when publish.
-          Def.task{
-            (new File(crossTarget.value, "classes")).mkdirs()
-            Seq.empty[File]
-          }.taskValue
-        }
-      ).settings(osgiSettings: _*).settings(
-        OsgiKeys.exportPackage := Seq(
-          "org.scalatest",
-          "org.scalatest.compatible",
-          "org.scalatest.concurrent",
-          "org.scalatest.diagrams",
-          "org.scalatest.enablers",
-          "org.scalatest.events",
-          "org.scalatest.exceptions",
-          "org.scalatest.expectations",
-          "org.scalatest.fixture",
-          "org.scalatest.funsuite",
-          "org.scalatest.featurespec",
-          "org.scalatest.funspec",
-          "org.scalatest.freespec",
-          "org.scalatest.flatspec",
-          "org.scalatest.matchers",
-          "org.scalatest.matchers.should",
-          "org.scalatest.matchers.must",
-          "org.scalatest.matchers.dsl",
-          "org.scalatest.verbs",
-          "org.scalatest.path",
-          "org.scalatest.prop",
-          "org.scalatest.propspec",
-          "org.scalatest.tags",
-          "org.scalatest.tagobjects",
-          "org.scalatest.time",
-          "org.scalatest.tools",
-          "org.scalatest.verb",
-          "org.scalatest.words",
-          "org.scalatest.wordspec",
-          "org.scalactic",
-          "org.scalactic.anyvals",
-          "org.scalactic.exceptions",
-          "org.scalactic.source"
-        ),
-        OsgiKeys.importPackage := Seq(
-          "org.scalatest.*",
-          "org.scalactic.*",
-          "scala.util.parsing.*;version=\"$<range;[==,=+);$<replace;1.0.4;-;.>>\"",
-          "scala.xml.*;version=\"$<range;[==,=+);$<replace;1.0.4;-;.>>\"",
-          "scala.*;version=\"$<range;[==,=+);$<replace;"+scalaBinaryVersion.value+";-;.>>\"",
-          "*;resolution:=optional"
-        ),
-        OsgiKeys.additionalHeaders:= Map(
-          "Bundle-Name" -> "ScalaTest",
-          "Bundle-Description" -> "ScalaTest is an open-source test framework for the Java Platform designed to increase your productivity by letting you write fewer lines of test code that more clearly reveal your intent.",
-          "Bundle-DocURL" -> "http://www.scalatest.org/",
-          "Bundle-Vendor" -> "Artima, Inc.",
-          "Main-Class" -> "org.scalatest.tools.Runner"
-        )
-      ).dependsOn(scalacticMacroNative % "compile-internal, test-internal", scalacticNative % "compile-internal", scalatestNative % "compile-internal").aggregate(scalacticMacroNative, scalacticNative, scalatestNative, commonTestNative, scalacticTestNative, scalatestTestNative).enablePlugins(ScalaNativePlugin)
 
   def gentestsLibraryDependencies =
     Seq(
